requirements.txt
abakit @ file:///home/eodonnell/programming/abakit
asgiref==3.8.1
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.11.1
boto3==1.26.100
botocore==1.29.100
brain-atlas-toolkit==0.4.3
Brotli==1.0.9
build==0.10.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==4.1.0
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cloud-files==4.12.1
cloud-volume==8.9.1
compressed-segmentation==2.2.0
compresso==3.2.0
configparser==5.3.0
connected-components-3d==3.10.5
constantly==15.1.0
contourpy==1.0.7
crc32c==2.3.post0
cryptography==39.0.0
cycler==0.11.0
daphne==2.5.0
dash==2.9.1
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
deflate==0.3.0
defusedxml==0.7.1
diff-match-patch==20200713
dijkstra3d==1.12.0
dill==0.3.6
dj-database-url==1.2.0
Django==5.0.7
django-allauth==0.57.0
django-compat==1.0.15
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-extensions==3.2.1
django-import-export==3.1.0
django-mysql==4.11.0
django-plotly-dash==2.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dpd-components==0.1.0
DracoPy==1.3.0
edt==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastremap==1.13.4
fill-voids==2.0.3
Flask==2.1.3
fonttools==4.39.2
fpzip==1.2.0
gevent==22.10.2
google-api-core==2.11.0
google-auth==2.16.3
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-logging==3.5.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
graphviz==0.20.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
igneous-pipeline==4.6.0
imageio==2.27.0
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.11
jsonschema==4.17.3
kimimaro==3.3.0
kiwisolver==1.4.4
lazy_loader==0.2
mapbuffer==0.6.0
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib==3.7.1
multiprocess==0.70.14
mysqlclient==2.1.1
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.7.0.72
openpyxl==3.1.1
orjson==3.8.8
packaging==23.0
pandas==1.5.3
pathos==0.3.0
pbr==5.11.1
pillow==10.3.0
plotly==5.13.1
pluggy==1.0.0
posix-ipc==1.1.1
pox==0.3.2
ppft==1.7.6.6
proto-plus==1.22.2
protobuf==4.22.1
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfqmr==0.1.2
Pygments==2.14.0
PyJWT==2.6.0
PyMCubes==0.1.4
PyMySQL==1.1.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pysimdjson==5.0.2
pyspng-seunglab==1.0.0
pytest==7.2.2
python-dateutil==2.8.2
python-decouple==3.8
python-jsonschema-objects==0.4.1
python3-openid==3.2.0
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-image==0.20.0
scipy==1.10.0
service-identity==21.1.0
SimpleITK==2.2.0
simplejpeg==1.6.5
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlparse==0.4.3
tablib==3.3.0
task-queue==2.12.0
tenacity==8.2.2
tifffile==2023.3.21
tinybrain==1.3.1
tomli==2.0.1
tqdm==4.65.0
trimesh==3.21.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
Unipath==1.1
urllib3==1.26.14
urllib3-secure-extra==0.1.0
vtk==9.2.6
Werkzeug==3.0.3
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
zfpc==0.1.2
zfpy==1.0.0
zmesh==1.6.2
zope.event==4.6
zope.interface==6.0
zstandard==0.20.0
